rename bot_controller to webhooks_controller
